refactor: new request methods for domain
