+++ src/renderer/domains/collectives/referendumMeta/types.ts
@@ -1,4 +1,6 @@
+import { type ChainId } from '@/shared/core';
 import { type ReferendumId } from '@/shared/pallet/referenda';
+import { type CollectivePalletsType } from '../_lib/types';
 
 export type ReferendumMetaProvider = 'subsquare' | 'polkassembly';
 
@@ -9,4 +11,6 @@ export type ReferendumMeta = {
   track: number;
   created: number;
   status: string;
+  pallet: CollectivePalletsType;
+  chainId: ChainId;
 };
